Added view controllers for each type of location and modified segue processing.
Processing for certain segues no longer is a direct result of pressing a button. It is not complete, but once it is, processing will be able to occur that allows for different segues and controllers to appear depending on the type of location that is held within a location cell.